fix banner
