+++ static/app/views/alerts/rules/metric/details/index.tsx
@@ -1,5 +1,4 @@
 import {Component, Fragment} from 'react';
-import styled from '@emotion/styled';
 import type {Location} from 'history';
 import isEqual from 'lodash/isEqual';
 import pick from 'lodash/pick';
@@ -8,12 +7,10 @@ import moment from 'moment-timezone';
 import {fetchOrgMembers} from 'sentry/actionCreators/members';
 import type {Client, ResponseMeta} from 'sentry/api';
 import {Alert} from 'sentry/components/core/alert';
-import {Button} from 'sentry/components/core/button';
 import {DateTime} from 'sentry/components/dateTime';
 import * as Layout from 'sentry/components/layouts/thirds';
 import PageFiltersContainer from 'sentry/components/organizations/pageFilters/container';
 import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
-import {IconClose} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import type {RouteComponentProps} from 'sentry/types/legacyReactRouter';
 import type {Organization} from 'sentry/types/organization';
@@ -33,8 +30,6 @@ import {
   fetchIncident,
   fetchIncidentsForRule,
 } from 'sentry/views/alerts/utils/apiCalls';
-import {DEPRECATED_TRANSACTION_ALERTS} from 'sentry/views/alerts/wizard/options';
-import {getAlertTypeFromAggregateDataset} from 'sentry/views/alerts/wizard/utils';
 
 import MetricDetailsBody from './body';
 import type {TimePeriodType} from './constants';
@@ -57,12 +52,16 @@ interface State {
   selectedIncident: Incident | null;
   incidents?: Incident[];
   rule?: MetricRule;
-  showTransactionsDeprecationAlert?: boolean;
   warning?: string;
 }
 
 class MetricAlertDetails extends Component<Props, State> {
-  state: State = {isLoading: false, hasError: false, error: null, selectedIncident: null};
+  state: State = {
+    isLoading: false,
+    hasError: false,
+    error: null,
+    selectedIncident: null,
+  };
 
   componentDidMount() {
     const {api, organization} = this.props;
@@ -70,7 +69,6 @@ class MetricAlertDetails extends Component<Props, State> {
     fetchOrgMembers(api, organization.slug);
     this.fetchData();
     this.trackView();
-    this.setState({showTransactionsDeprecationAlert: true});
   }
 
   componentDidUpdate(prevProps: Props) {
@@ -261,14 +259,7 @@ class MetricAlertDetails extends Component<Props, State> {
   }
 
   render() {
-    const {
-      rule,
-      incidents,
-      hasError,
-      selectedIncident,
-      warning,
-      showTransactionsDeprecationAlert,
-    } = this.state;
+    const {rule, incidents, hasError, selectedIncident, warning} = this.state;
     const {organization, projects, loadingProjects} = this.props;
     const timePeriod = this.getTimePeriod(selectedIncident);
 
@@ -278,18 +269,6 @@ class MetricAlertDetails extends Component<Props, State> {
 
     const project = projects.find(({slug}) => slug === rule?.projects[0]);
     const isGlobalSelectionReady = project !== undefined && !loadingProjects;
-
-    const ruleType =
-      rule &&
-      getAlertTypeFromAggregateDataset({
-        aggregate: rule.aggregate,
-        dataset: rule.dataset,
-        eventTypes: rule.eventTypes,
-        organization,
-      });
-
-    const deprecateTransactionsAlertsWarning =
-      ruleType && DEPRECATED_TRANSACTION_ALERTS.includes(ruleType);
 
     return (
       <PageFiltersContainer
@@ -312,29 +291,6 @@ class MetricAlertDetails extends Component<Props, State> {
           project={project}
           onSnooze={this.onSnooze}
         />
-        {showTransactionsDeprecationAlert && deprecateTransactionsAlertsWarning && (
-          <Alert.Container>
-            <Alert
-              type="warning"
-              style={{marginBottom: 0}}
-              trailingItems={
-                <StyledCloseButton
-                  icon={<IconClose size="sm" />}
-                  aria-label={t('Close')}
-                  onClick={() => {
-                    this.setState({showTransactionsDeprecationAlert: false});
-                  }}
-                  size="zero"
-                  borderless
-                />
-              }
-            >
-              {t(
-                'Transaction based alerts are going to be deprecated soon. Please use Span alerts instead.'
-              )}
-            </Alert>
-          </Alert.Container>
-        )}
         <MetricDetailsBody
           rule={rule}
           project={project}
@@ -348,14 +304,3 @@ class MetricAlertDetails extends Component<Props, State> {
 }
 
 export default withApi(withProjects(MetricAlertDetails));
-
-const StyledCloseButton = styled(Button)`
-  background-color: transparent;
-  transition: opacity 0.1s linear;
-
-  &:hover,
-  &:focus {
-    background-color: transparent;
-    opacity: 1;
-  }
-`;
