+++ static/app/views/alerts/rules/metric/ruleConditionsForm.tsx
@@ -33,7 +33,7 @@ import Panel from 'sentry/components/panels/panel';
 import PanelBody from 'sentry/components/panels/panelBody';
 import {SearchQueryBuilder} from 'sentry/components/searchQueryBuilder';
 import {InvalidReason} from 'sentry/components/searchSyntax/parser';
-import {t, tct} from 'sentry/locale';
+import {t, tct, tctCode} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {SelectValue} from 'sentry/types/core';
 import type {Tag, TagCollection} from 'sentry/types/group';
@@ -559,8 +559,8 @@ class RuleConditionsForm extends PureComponent<Props, State> {
         {deprecateTransactionsAlertsWarning && (
           <Alert.Container>
             <Alert type="warning">
-              {t(
-                'Transaction based alerts are going to be deprecated soon. Please use Span alerts instead.'
+              {tctCode(
+                'Transaction based alerts are going to be deprecated soon. Please use Span alerts instead. You can isolate transactions by using the [code:is_transaction:true] filter.'
               )}
             </Alert>
           </Alert.Container>
